refactor: 새로운 폴더 구조에 맞춰서 폴더 변경
