+++ breaking-front/src/pages/Profile/components/ProfileFollowModal/ProfileFollowModal.js
@@ -2,8 +2,8 @@ import FollowCard from 'components/FollowCard/FollowCard';
 import Modal from 'components/Modal/Modal';
 import { FollowCardSkeleton } from 'components/Skeleton/Skeleton';
 import { PAGE_PATH } from 'constants/path';
-import useFollowerList from 'hooks/queries/useFollowerList';
-import useFollowingList from 'hooks/queries/useFollowingList';
+import useFollowerList from 'pages/Profile/hooks/useFollowerList';
+import useFollowingList from 'pages/Profile/hooks/useFollowingList';
 import React, { useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
